requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.27.0
Babel==2.9.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
Eel==0.14.0
exceptiongroup==1.1.0
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pefile==2022.5.30
Pillow==9.4.0
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pywin32-ctypes==0.2.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
whichcraft==0.6.1
wsproto==1.2.0
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
